adjusts to client and admin route sections for layout purposes
