+++ app/(admin)/admin/page.tsx
@@ -1,0 +1,3 @@
+export default function Home() {
+	return <div>hi</div>;
+}
