+++ app/(client)/@modal/(.)gallery/[id]/ModalWrapper.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import AnimateWrapper from "@/app/components/components/AnimateWrapper";
-import Card from "@/app/components/components/Card";
+import AnimateWrapper from "@/app/(client)/components/components/AnimateWrapper";
+import Card from "@/app/(client)/components/components/Card";
 import { motion } from "framer-motion";
 import { useRouter } from "next/navigation";
 const ModalWrapper = ({ children }: { children: React.ReactNode }) => {
