+++ app/(client)/components/SVGs/EmailLogo.tsx
