+++ app/(client)/components/SVGs/FullLogo.tsx
