+++ app/(client)/components/SVGs/IGLogo.tsx
