+++ app/(client)/components/SVGs/SmallLogo.tsx
