+++ app/(client)/components/layout/Footer.tsx
