+++ app/(client)/components/transitions/CoverPageTransition.tsx
