+++ app/(client)/gallery/components/GallerySliderWrapper.tsx
