+++ app/(client)/gallery/page.tsx
