+++ app/(client)/welcome/components/WelcomeDisplay.tsx
