+++ app/(client)/welcome/components/WelcomeImage.tsx
