+++ app/(client)/welcome/page.tsx
