+++ app/utilities/resendFunctions.ts
@@ -2,9 +2,9 @@
 
 import { ReactElement } from "react";
 import { Resend } from "resend";
-import { CustomerEmailTemplate } from "../contact/components/CustomerEmailTemplate";
-import { InboxEmailTemplate } from "../contact/components/InboxEmailTemplate";
 import { ContactFormData } from "./types";
+import { CustomerEmailTemplate } from "../(client)/contact/components/CustomerEmailTemplate";
+import { InboxEmailTemplate } from "../(client)/contact/components/InboxEmailTemplate";
 
 export const sendCustomerEmail = async (contactFormData: ContactFormData) => {
 	const resend = new Resend(process.env.RESEND_API_KEY);
